+++ .vscode/configurationCache.log
@@ -1,1 +1,1 @@
-{"buildTargets":[],"launchTargets":["/media/kaelan/Disk/Dev/TCPforwarderControl/build>TCPforward()"],"customConfigurationProvider":{"workspaceBrowse":{"browsePath":["/media/kaelan/Disk/Dev/TCPforwarderControl/src"],"compilerArgs":["-pthread","-lm","-ggdb","-o","build/TCPforward"],"compilerPath":"/usr/local/bin/g++","windowsSdkVersion":""},"fileIndex":[["/media/kaelan/Disk/Dev/TCPforwarderControl/src/TCPforward.cpp",{"uri":{"$mid":1,"fsPath":"/media/kaelan/Disk/Dev/TCPforwarderControl/src/TCPforward.cpp","path":"/media/kaelan/Disk/Dev/TCPforwarderControl/src/TCPforward.cpp","scheme":"file"},"configuration":{"defines":[],"includePath":[],"forcedInclude":[],"compilerPath":"/usr/local/bin/g++","compilerArgs":["-pthread","-lm","-ggdb","-o","build/TCPforward"],"windowsSdkVersion":""},"compileCommand":{"command":"g++ src/TCPforward.cpp -pthread -lm -ggdb -o build/TCPforward","directory":"/media/kaelan/Disk/Dev/TCPforwarderControl","file":"/media/kaelan/Disk/Dev/TCPforwarderControl/src/TCPforward.cpp"}}]]}}
+{"buildTargets":["all","bin","clean","deb","install"],"launchTargets":["/media/kaelan/Disk/Dev/TCPforwarderControl/build>TCPforward()"],"customConfigurationProvider":{"workspaceBrowse":{"browsePath":["/media/kaelan/Disk/Dev/TCPforwarderControl/src"],"compilerArgs":["-pthread","-lm","-ggdb","-o","build/TCPforward"],"compilerPath":"/usr/local/bin/g++","windowsSdkVersion":""},"fileIndex":[["/media/kaelan/Disk/Dev/TCPforwarderControl/src/TCPforward.cpp",{"uri":{"$mid":1,"fsPath":"/media/kaelan/Disk/Dev/TCPforwarderControl/src/TCPforward.cpp","external":"file:///media/kaelan/Disk/Dev/TCPforwarderControl/src/TCPforward.cpp","path":"/media/kaelan/Disk/Dev/TCPforwarderControl/src/TCPforward.cpp","scheme":"file"},"configuration":{"defines":[],"includePath":[],"forcedInclude":[],"compilerPath":"/usr/local/bin/g++","compilerArgs":["-pthread","-lm","-ggdb","-o","build/TCPforward"],"windowsSdkVersion":""},"compileCommand":{"command":"g++ src/TCPforward.cpp -pthread -lm -ggdb -o build/TCPforward","directory":"/media/kaelan/Disk/Dev/TCPforwarderControl","file":"/media/kaelan/Disk/Dev/TCPforwarderControl/src/TCPforward.cpp"}}]]}}
