+++ src/TCPforward.cpp
@@ -1,15 +1,17 @@
 #include <iostream>
 #include "TCPstreams2.h"
-#include "StringInputs2.h"
+//#include "StringInputs2.h"
 #include <vector>
 #include <stdio.h>
 #include <pthread.h>
 #include <stdlib.h>
 #include <cstdlib>
 #include <signal.h>
-#include "LoadConfig2.h"
+//#include "LoadConfig2.h"
 #include "UNIXstreams.h"
 #include <math.h>
+#include "HelpMenu.h"
+#include "MainConfigLoader.h"
 
 using namespace std;
 
@@ -217,145 +219,9 @@ void garbageCollector(){
 
 }
 
-//Prints the Help Text
-void printHelp(char** argv){
-    printf("Usage:\n%s\n",argv[0]);
-    printf("\t-h\t: This Help Menu\n");
-    printf("\t-c\t: Configuration File, Specified by -c FilePath. EG: -c Example.conf\n");
-    printf("\t-host\t: The IP Address to host the TCP server on. Usually 0.0.0.0 (Optional)\n");
-    printf("\t-dest\t: The IP Address Of the destination.\n");
-    printf("\t-hport\t: The Hosting Port\n");
-    printf("\t-dport\t: The Destination Port\n");
-    printf("\t-buf\t: The Buffer size, usually 1024 (Optional)\n");
-	printf("\t-climit\t: The Connection Limit, Set to 0 for unlimited. Default is 0 (Optional)\n");
-	printf("\t-control\t: The Control Unix Socket (Optional) [Work in Progress]\n");
-}
 
 
 
-//Required Configuration Values
-const int confValues = 3;
-
-//For the File
-const char* ConfTags[3] {
-    "Client_address",
-    "Client_port",
-    "Host_port",
-
-};
-
-//For the Cmdline Arguments
-const char* ArgTags[3]{
-    "-dest",
-    "-hport",
-    "-dport",
-};
-
-//This function loads the Configuration
-int loadSettings(int argc, char** argv){
-    struct configData* MyArgs = readArgs(argc,argv);
-    //printf("Got the Following Arguments:\n");
-    //printConfig(MyArgs);
-    if (getConfValue("-c",MyArgs) != NULL){
-        char* CF_Name = getConfValue("-c",MyArgs);
-        printf("Config File Specified, %s\n",CF_Name);
-
-        //DeAllocate Args
-        destroyArgsConfig(MyArgs);
-
-        //Load the Config into the same struct pointer
-        MyArgs = readConfig(CF_Name);
-
-        //Check Flags
-        for (int i = 0; i < confValues; i++){
-            //printf("Loading %s\n",ConfTags[i]);
-            if (getConfValue(ConfTags[i],MyArgs) == NULL){
-                printf("%s is not defined\n",ConfTags[i]);
-                destroyConfig(MyArgs);
-                return -1;
-            }
-        }
-
-        //Assign Values
-        DestIP.assign(getConfValue("Client_address",MyArgs));
-        HostPort = stosi(getConfValue("Host_port",MyArgs));
-        DestPort = stosi(getConfValue("Client_port",MyArgs));
-        
-
-       
-        //Optional Arguments
-        if (getConfValue("Host_address",MyArgs) != NULL){
-            HostIP.assign(getConfValue("Host_address",MyArgs));
-        }
-        if (getConfValue("Buffer_size",MyArgs) != NULL){
-            bufsize = stosi(getConfValue("Buffer_size",MyArgs));
-        }
-        if (getConfValue("Connection_limit",MyArgs) != NULL){
-            ConLimit = stosi(getConfValue("Connection_limit",MyArgs));
-        }
-        if (getConfValue("Control_socket",MyArgs) != NULL){
-            ControlSocketPath.assign(getConfValue("Control_socket",MyArgs));
-            ControlSocket = 1;
-        }
-
-        
-        destroyConfig(MyArgs);
-
-        return 1;
-
-    }
-
-    //Help Command Specified
-    else if (getConfValue("-h",MyArgs) != NULL){
-        destroyArgsConfig(MyArgs);
-        return -2;
-    }
-    
-    
-    //Loads CMDline Arguments
-    else {
-        //Check Flags
-        for (int i = 0; i < confValues; i++){
-            //printf("Loading %s\n",ArgTags[i]);
-            if (getConfValue(ArgTags[i],MyArgs) == NULL){
-                printf("%s Required Argument not Provided!\n",ArgTags[i]);
-                destroyArgsConfig(MyArgs);
-                return -1;
-            }
-        }
-
-
-        //Assign Arguments            
-        DestIP.assign(getConfValue("-dest",MyArgs));
-        HostPort = stosi(getConfValue("-hport",MyArgs));
-        DestPort = stosi(getConfValue("-dport",MyArgs));
-
-
-        //Optional Arguments
-        if (getConfValue("-host",MyArgs) != NULL){
-            HostIP.assign(getConfValue("-host",MyArgs));
-        }
-
-        if (getConfValue("-buf",MyArgs) != NULL){
-            bufsize = stosi(getConfValue("-buf",MyArgs));
-        }
-        if (getConfValue("-climit",MyArgs) != NULL){
-            ConLimit = stosi(getConfValue("-climit",MyArgs));
-        }
-        if (getConfValue("-control",MyArgs) != NULL){
-            ControlSocketPath.assign(getConfValue("-control",MyArgs));
-            ControlSocket = 1;
-        }
-
-        destroyArgsConfig(MyArgs);
-
-        return 0;
-
-    }
-
-
-    
-}
 
 
 //This is the main server socket for 
@@ -620,7 +486,7 @@ int main(int argc, char** argv){
         return -1; 
     }
 
-    int loadResult = loadSettings(argc,argv);
+    int loadResult = loadSettings(argc,argv,&HostIP,&DestIP,&HostPort,&DestPort,&bufsize,&ControlSocketPath,&ControlSocket,&ConLimit);
     if (loadResult < 0){
         printHelp(argv);
         return -1;
